+++ src/DependencyInjection/SentryExtension.php
@@ -32,7 +32,7 @@ class SentryExtension extends Extension
      *
      * @throws InvalidConfigurationException
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $configuration = new Configuration();
         $processedConfiguration = $this->processConfiguration($configuration, $configs);