+++ src/DependencyInjection/Configuration.php
@@ -57,6 +57,11 @@ class Configuration implements ConfigurationInterface
         if (PrettyVersions::getVersion('sentry/sentry')->getPrettyVersion() !== '2.0.0') {
             $optionsChildNodes->booleanNode('capture_silenced_errors');
         }
+        if ($this->classSerializersAreSupported()) {
+            $optionsChildNodes->arrayNode('class_serializers')
+                ->defaultValue([])
+                ->prototype('scalar');
+        }
         $optionsChildNodes->integerNode('context_lines')
             ->min(0)
             ->max(99);
@@ -80,13 +85,13 @@ class Configuration implements ConfigurationInterface
             ->prototype('scalar')
             ->validate()
             ->ifTrue(function ($value): bool {
-                if (! is_string($value)) {
+                if (! is_string($value) && '' != $value) {
                     return true;
                 }
 
-                return '@' !== substr($value, 0, 1);
+                return '@' !== $value[0];
             })
-            ->thenInvalid('Expecting service reference, got %s');
+            ->thenInvalid('Expecting service reference, got "%s"');
         $optionsChildNodes->scalarNode('logger');
         $optionsChildNodes->integerNode('max_breadcrumbs')
             ->min(1);
@@ -162,5 +167,16 @@ class Configuration implements ConfigurationInterface
 
             return true;
         };
+    }
+
+    private function classSerializersAreSupported(): bool
+    {
+        try {
+            new Options(['class_serializers' => []]);
+
+            return true;
+        } catch (\Throwable $throwable) {
+            return false;
+        }
     }
 }